Fix warning.

+++ src/eepp/ui/doc/documentview.cpp
@@ -402,8 +402,8 @@ double DocumentView::getLineYOffset( Int64 docIdx, Float lineHeight ) const {
 
 bool DocumentView::isLineVisible( Int64 docIdx ) const {
 	return mDocLineToVisibleIndex.empty() ||
-		   ( docIdx < mDocLineToVisibleIndex.size() && mDocLineToVisibleIndex[docIdx] !=
-			 static_cast<Int64>( VisibleIndex::invalid ) );
+		   ( docIdx < static_cast<Int64>( mDocLineToVisibleIndex.size() ) &&
+			 mDocLineToVisibleIndex[docIdx] != static_cast<Int64>( VisibleIndex::invalid ) );
 }
 
 std::vector<TextRange> DocumentView::intersectsFoldedRegions( const TextRange& range ) const {
